Complete delete tracker logic
